requirements.txt
discord.py
groq
google-generativeai
requests
Pillow
